Lots of changes in login procedure

+++ PokemonGo.RocketAPI.Console/Program.cs
@@ -32,10 +32,7 @@ namespace PokeMaster
         public static string deviceData = Path.Combine(path_device, "DeviceData.json");
         public static string cmdCoords = string.Empty;
         public static string accountProfiles = Path.Combine(path, "Profiles.txt");
-        static readonly string logPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Logs");
-        public static string pokelog = Path.Combine(logPath, "PokeLog.txt");
-        public static string manualTransferLog = Path.Combine(logPath, "TransferLog.txt");
-        public static string EvolveLog = Path.Combine(logPath, "EvolveLog.txt");
+        public static string logPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Logs");
         public static string path_pokedata = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "PokeData");
 
         static Version version;       
@@ -116,6 +113,7 @@ namespace PokeMaster
                             }
                         }
 
+                        Logger.ColoredConsoleWrite(ConsoleColor.Red, "Using Profile: " + GlobalVars.ProfileName);
                         if (GlobalVars.UsePwdEncryption) GlobalVars.Password = Encryption.Decrypt(GlobalVars.Password);
                         #endregion
                     }
@@ -212,7 +210,8 @@ namespace PokeMaster
             }
             
             SleepHelper.PreventSleep();
-            CreateLogDirectories();
+
+            CreateLogDirectories(Path.Combine(logPath, GlobalVars.ProfileName));
 
             GlobalVars.infoObservable.HandleNewHuntStats += SaveHuntStats;
 
@@ -312,24 +311,16 @@ namespace PokeMaster
                         "https://raw.githubusercontent.com/Logxn/PokemonGo-Bot/master/ver.md");
         }
 
-        private static void CreateLogDirectories()
+        private static void CreateLogDirectories(string logPath)
         {
-            if (!Directory.Exists(logPath))
-            {
-                Directory.CreateDirectory(logPath);
-            }
-            if (!File.Exists(pokelog))
-            {
-                File.Create(pokelog).Close();
-            }
-            if (!File.Exists(manualTransferLog))
-            {
-                File.Create(manualTransferLog).Close();
-            }
-            if (!File.Exists(EvolveLog))
-            {
-                File.Create(EvolveLog).Close();
-            }
+            string pokelog = Path.Combine(logPath, "PokeLog.txt");
+            string manualTransferLog = Path.Combine(logPath, "TransferLog.txt");
+            string EvolveLog = Path.Combine(logPath, "EvolveLog.txt");
+
+            if (!Directory.Exists(logPath)) Directory.CreateDirectory(logPath);
+            if (!File.Exists(pokelog)) File.Create(pokelog).Close();
+            if (!File.Exists(manualTransferLog)) File.Create(manualTransferLog).Close();
+            if (!File.Exists(EvolveLog)) File.Create(EvolveLog).Close();
         }
     }
 }
